requirements/prod.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/prod.in
#
alembic==1.6.5
    # via flask-db
amqp==5.0.6
    # via kombu
appdirs==1.4.4
    # via black
attrs==21.2.0
    # via pytest
babel==2.9.1
    # via flask-babelex
bcrypt==3.2.0
    # via flask-user
billiard==3.6.4.0
    # via celery
black==21.6b0
    # via -r requirements/_core.in
blinker==1.4
    # via flask-mail
celery==5.1.0
    # via -r requirements/_core.in
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via bcrypt
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
click==7.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
    #   pip-tools
coverage==5.5
    # via pytest-cov
dnspython==2.1.0
    # via email-validator
email-validator==1.1.3
    # via -r requirements/_core.in
et-xmlfile==1.1.0
    # via openpyxl
flake8==3.9.2
    # via -r requirements/_core.in
flask-babelex==0.9.4
    # via -r requirements/_core.in
flask-db==0.3.2
    # via -r requirements/_core.in
flask-login==0.5.0
    # via
    #   -r requirements/_core.in
    #   flask-user
flask-mail==0.9.1
    # via flask-user
flask-pydantic==0.8.0
    # via -r requirements/_core.in
flask-serialize==1.5.1
    # via -r requirements/_core.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements/_core.in
    #   flask-db
    #   flask-user
    #   pytest-flask-sqlalchemy
flask-user==0.6.21
    # via -r requirements/_core.in
flask-wtf==0.15.1
    # via
    #   -r requirements/_core.in
    #   flask-user
flask==2.0.1
    # via
    #   -r requirements/_core.in
    #   flask-babelex
    #   flask-db
    #   flask-login
    #   flask-mail
    #   flask-pydantic
    #   flask-sqlalchemy
    #   flask-user
    #   flask-wtf
gunicorn==20.1.0
    # via -r requirements/_core.in
idna==2.10
    # via
    #   email-validator
    #   requests
iniconfig==1.1.1
    # via pytest
itsdangerous==2.0.1
    # via
    #   flask
    #   flask-wtf
jinja2==3.0.1
    # via
    #   flask
    #   flask-babelex
kombu==5.1.0
    # via celery
mako==1.1.4
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
    #   wtforms
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
numexpr==2.7.3
    # via tables
numpy==1.20.3
    # via
    #   numexpr
    #   pandas
    #   tables
openpyxl==3.0.7
    # via -r requirements/_core.in
packaging==20.9
    # via
    #   pytest
    #   pytest-flask-sqlalchemy
pandas==1.2.4
    # via -r requirements/_core.in
passlib==1.7.4
    # via flask-user
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
permissive-dict==1.0.4
    # via flask-serialize
pip-tools==6.1.0
    # via -r requirements/_core.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.18
    # via click-repl
psycopg2==2.8.6
    # via -r requirements/prod.in
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pycryptodome==3.10.1
    # via flask-user
pydantic-sqlalchemy==0.0.9
    # via -r requirements/_core.in
pydantic==1.8.2
    # via
    #   -r requirements/_core.in
    #   flask-pydantic
    #   pydantic-sqlalchemy
pyflakes==2.3.1
    # via flake8
pyparsing==2.4.7
    # via packaging
pytest-cov==2.12.1
    # via -r requirements/_core.in
pytest-env==0.6.2
    # via -r requirements/_core.in
pytest-flask-sqlalchemy==1.0.2
    # via -r requirements/_core.in
pytest-mock==3.6.1
    # via pytest-flask-sqlalchemy
pytest==6.2.4
    # via
    #   -r requirements/_core.in
    #   pytest-cov
    #   pytest-env
    #   pytest-flask-sqlalchemy
    #   pytest-mock
python-dateutil==2.8.1
    # via
    #   alembic
    #   pandas
python-dotenv==0.17.1
    # via -r requirements/_core.in
python-editor==1.0.4
    # via alembic
pytz==2021.1
    # via
    #   babel
    #   celery
    #   pandas
pyyaml==5.4.1
    # via -r requirements/_core.in
regex==2021.4.4
    # via black
requests==2.25.1
    # via -r requirements/_core.in
six==1.16.0
    # via
    #   bcrypt
    #   click-repl
    #   python-dateutil
    #   sqlalchemy-utils
speaklater==1.3
    # via flask-babelex
sqlalchemy-utils==0.37.7
    # via flask-db
sqlalchemy==1.3.23
    # via
    #   -r requirements/_core.in
    #   alembic
    #   flask-db
    #   flask-sqlalchemy
    #   pydantic-sqlalchemy
    #   pytest-flask-sqlalchemy
    #   sqlalchemy-utils
tables==3.6.1
    # via -r requirements/_core.in
toml==0.10.2
    # via
    #   black
    #   pep517
    #   pytest
    #   pytest-cov
typing-extensions==3.10.0.0
    # via pydantic
urllib3==1.26.5
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.1
    # via flask
wtforms==2.3.3
    # via
    #   -r requirements/_core.in
    #   flask-wtf
xlsxwriter==3.0.1
    # via -r requirements/_core.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
